@@ -14,7 +14,7 @@
       <li>...</li>
     </ul>
   </div>
-  <button on:click={logout} class="btn btn-primary">Uitloggen</button>
+  <button onclick={logout} class="btn btn-primary">Uitloggen</button>
 {:else}
   <p>
     Je bent momenteel niet ingelogd. Druk op de knop hieronder om naar de login
